backend/requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
